RSA Base complete, need to implement into GUI + finish key methods
I don't know exactly how you want to use format keys, so I'll leave the implementation up to you. The important information to store should be commented in the method body.

Currently, it uses fixed values for p, q, and e. If you want, we could have these be updated. It would need a method, but the declarations/everything else are the same. If you have questions, I'll probably be on discord most of the time.